+++ json_serializable/lib/src/json_serializable_generator.dart
@@ -12,7 +12,7 @@ import 'encoder_helper.dart';
 import 'field_helpers.dart';
 import 'helper_core.dart';
 import 'type_helper.dart';
-import 'type_helper_context.dart';
+import 'type_helper_ctx.dart';
 import 'type_helpers/convert_helper.dart';
 import 'type_helpers/date_time_helper.dart';
 import 'type_helpers/enum_helper.dart';